make sparse.py
